Address playground loopback review feedback
